feat: list의 비어있음 여부 확인

+++ backend/src/main/java/com/shyashyashya/refit/domain/qnaset/service/QnaSetService.java
@@ -68,7 +68,9 @@ public class QnaSetService {
 
     private void deleteAllHighlightingsAndRects(QnaSet qnaSet) {
         List<Long> pdfHighlightingIds = pdfHighlightingRepository.findIdsByQnaSet(qnaSet);
-        pdfHighlightingRectRepository.deleteAllByPdfHighlightingIds(pdfHighlightingIds);
+        if (!pdfHighlightingIds.isEmpty()) {
+            pdfHighlightingRectRepository.deleteAllByPdfHighlightingIds(pdfHighlightingIds);
+        }
         pdfHighlightingRepository.deleteAllByQnaSet(qnaSet);
     }
